fix raylib2v header file conversion and update fncs

+++ mouse.v
@@ -78,6 +78,6 @@ pub fn set_mouse_cursor(cursor int) {
 }
 
 [inline]
-pub fn get_mouse_wheel_move() int {
+pub fn get_mouse_wheel_move() f32 {
 	return C.GetMouseWheelMove()
 }
